Base implementation of service to clear database.
